CIP: Combine handling for cip_connID_info_t connections
Notes:
1. There are no functionality changes with this delivery
2. cip_connID_info_t describes a one-way connection. Each CIP Connection includes 2 of these. Previously, each operation was duplicated for each direction.
3. This change is to reduce copypaste, simplify logic, and make it easier to add upcoming feature changes, and fix some connection handling issues (future merge requests)

Changes:
1. Extract Method: get_conversation_info_one_direction
2. dissect_net_param16/dissect_net_param32: Parse and set data into cip_connID_info_t

+++ epan/dissectors/packet-cip.h
@@ -469,11 +469,12 @@ typedef struct attribute_info {
    attribute_dissector_func *pdissect;
 } attribute_info_t;
 
+// This describes a one-way connection. Each CIP Connection includes 2 of these.
 typedef struct cip_connID_info {
    guint32 connID;
    address ipaddress;
    guint16 port;
-   guint8  type;
+   guint32 type;
 
    // Actual Packet Interval in microseconds.
    guint32 api;
